requirements.txt
requests
openpyxl
slack_sdk
tenacity
python-dotenv
aiohttp
